Extract base class [#121638271]
In preparation for introducing parallel append, we extracted a base
class from (formerly) CPhysicalUnionAll. To the less familiar in the
audience, here's a before-after:

- Before:
  0. `gpopt::CPhysicalUnionAll`

- After:
  0. `gpopt::CPhysicalUnionAll`
  0. `gpopt::CPhysicalSerialUnionAll` is now a subclass of the
  abstract base class `gpopt::CPhysicalUnionAll`

This makes it much less difficult to introduce
`CPhysicalParallelUnionAll` as a subclass of `CPhysicalUnionAll`.
